reconnect to master_dbh as needed
The problem this is really meant to solve is that during global
destruction, the master_dbh may be destroyed before some given
tempdb.  TempDB then tries to cleanup through master dbh, but it
has gone away.  This will re-connect as needed, since it is
surely saner than trying to order global destruction!